move ttl client to thriftclient package
